+++ src/components/scene/DynamicObject.tsx
@@ -4,6 +4,7 @@ import TorusKnotObject from './objects/TorusKnotObject';
 import WobbleFieldObject from './objects/WobbleFieldObject';
 import DistortionSphereObject from './objects/DistortionSphereObject';
 import MorphingIcosahedronObject from './objects/MorphingIcosahedronObject';
+import WavyGridObject from './objects/WavyGridObject';
 
 interface DynamicObjectProps {
   type: SceneConfig['type'];
@@ -21,6 +22,8 @@ const DynamicObject = ({ type, themeConfig }: DynamicObjectProps) => {
       return <DistortionSphereObject color={mainObjectColor} materialConfig={material} />;
     case 'MorphingIcosahedron':
       return <MorphingIcosahedronObject color={mainObjectColor} materialConfig={material} />;
+    case 'WavyGrid':
+      return <WavyGridObject color={mainObjectColor} materialConfig={material} />;
     default:
       return null;
   }